Refines project structure and simplifies some elements
Removes the system manager in favor of Globals.h a file with global
variables, containing the systems needed in different parts of the engine.

Everything having to do with graphics has been moved into it's own
directory Graphics, and the code has been wrapped in a 'Graphics'
namespace.

Seeing no need for the Matrix and Vector classes to be templates, that is
I only see the use of 4x4 matrices, and vectors of size 2, 3 or 4, these
classes has been non-templateized, and separated.

Most of the functions are still inlined, including operators.

This updates breaks the build, as there are some external references that
are not correct. Hopefully this an easy fix.